+++ models/User.js
@@ -3,7 +3,7 @@
 
 const { Model, DataTypes } = require('sequelize');
 const sequelize = require('../config/connection');
-
+const bcrypt = require('bcrypt');
 //creates model
 class User extends Model {
     checkPassword(loginPw){
@@ -48,6 +48,18 @@ User.init(
         }
     },
     {
+        hooks: {
+            // set up beforeCreate lifecycle "hook" functionality
+        async beforeCreate(newData) {
+            newData.password = await bcrypt.hash(newData.password, 10);
+               return newData;
+             
+           },
+           async beforeUpdate(updatedData){
+             updatedData.password = await bcrypt.hash(updatedData.password, 10);
+             return updatedData; 
+           }
+        },
         //Defines table config opts
         sequelize,
         timestamps: false,
