+++ package.json
@@ -20,6 +20,7 @@
   "homepage": "https://github.com/Narmstrongv/HTeamL-Project-2#readme",
   "dependencies": {
     "bcrpyt": "^2.0.0",
+    "bcrypt": "^5.0.1",
     "connect-session-sequelize": "^7.1.1",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
